Extract compound components and utilities from large Svelte files
Break apart several oversized components into focused, reusable pieces:

- Extract AbsorbPlanModal, DiscardChangesModal, and StashIntoBranchModal
  from ChangedFilesContextMenu
- Extract RuleEditor from RulesList
- Extract UpstreamIntegrationActions from BranchCommitList
- Extract pure step manipulation functions from BranchIntegrationModal
  into integrationStepUtils

Also fix Modal Enter key handling: when focus is not on a native input
or button, pressing Enter now clicks the submit button (preserving
AsyncButton loading state) or falls back to calling onSubmit directly.
