+++ apps/desktop/src/components/RulesList.svelte
@@ -1,38 +1,19 @@
 <script lang="ts">
 	import Drawer from "$components/Drawer.svelte";
-	import NewRuleMenu from "$components/NewRuleMenu.svelte";
 	import ReduxResult from "$components/ReduxResult.svelte";
 	import Rule from "$components/Rule.svelte";
-	import RuleFiltersEditor from "$components/RuleFiltersEditor.svelte";
+	import RuleEditor from "$components/RuleEditor.svelte";
 	import {
 		type WorkspaceRuleId,
 		type RuleFilterMap,
-		type RuleFilterType,
-		type WorkspaceRule,
 		type StackTarget,
-		encodeStackTarget,
-		decodeStackTarget,
-		compareStackTarget,
+		type WorkspaceRule,
 		type RuleFilter,
 	} from "$lib/rules/rule";
 	import { RULES_SERVICE, workspaceRulesSelectors } from "$lib/rules/rulesService.svelte";
-	import { getStackName } from "$lib/stacks/stack";
-	import { STACK_SERVICE } from "$lib/stacks/stackService.svelte";
-	import { typedKeys } from "$lib/utils/object";
 	import { inject } from "@gitbutler/core/context";
-	import {
-		Button,
-		chipToasts,
-		Select,
-		SelectItem,
-		Icon,
-		Badge,
-		Spacer,
-		SkeletonBone,
-		Link,
-	} from "@gitbutler/ui";
+	import { Badge, Button, chipToasts, Link, SkeletonBone } from "@gitbutler/ui";
 	import { focusable } from "@gitbutler/ui/focus/focusable";
-	import { isDefined } from "@gitbutler/ui/utils/typeguards";
 	import { slide } from "svelte/transition";
 
 	type Props = {
@@ -42,53 +23,20 @@
 	const { projectId }: Props = $props();
 
 	const rulesService = inject(RULES_SERVICE);
-	const stackService = inject(STACK_SERVICE);
-
-	const [create, creatingRule] = rulesService.createWorkspaceRule;
-	const [update, updatingRule] = rulesService.updateWorkspaceRule;
-
-	let selectedRuleId = $state<WorkspaceRuleId>();
-	let stackTargetSelected = $state<StackTarget>();
-	const encodedStackTarget = $derived(
-		stackTargetSelected && encodeStackTarget(stackTargetSelected),
-	);
-
-	let draftRuleFilterInitialValues = $state<Partial<RuleFilterMap>>({});
-	const ruleFilterTypes = $derived(typedKeys(draftRuleFilterInitialValues));
-
-	// Component references
-	let ruleFiltersEditor = $state<RuleFiltersEditor>();
-
-	let addFilterButton = $state<HTMLDivElement>();
-	let newFilterContextMenu = $state<NewRuleMenu>();
 
 	// Visual state
 	let mode = $state<"list" | "edit" | "add">("list");
 	let editingRuleId = $state<WorkspaceRuleId | null>(null);
 
+	// Initial values passed to RuleEditor when opening edit mode
+	let editorInitialStackTarget = $state<StackTarget | undefined>(undefined);
+	let editorInitialFilterValues = $state<Partial<RuleFilterMap>>({});
+
 	// Read initial collapsed state from persisted storage, default to false (open) if not found
 	const drawerPersistId = `rules-drawer-${projectId}`;
 	let drawer = $state<Drawer>();
 
-	const validFilters = $derived(!ruleFiltersEditor || ruleFiltersEditor.imports.filtersValid);
-	const canSaveRule = $derived(stackTargetSelected !== undefined && validFilters);
-
-	function openAddFilterContextMenu(e: MouseEvent) {
-		e.stopPropagation();
-		newFilterContextMenu?.toggle(e);
-	}
-
-	function addDraftRuleFilter(type: RuleFilterType) {
-		const ruleTypes = typedKeys(draftRuleFilterInitialValues);
-		if (!ruleTypes.includes(type)) {
-			draftRuleFilterInitialValues[type] = null;
-		}
-	}
-
-	function removeDraftRuleFilter(type: RuleFilterType) {
-		draftRuleFilterInitialValues[type] = undefined;
-		delete draftRuleFilterInitialValues[type];
-	}
+	const rules = $derived(rulesService.workspaceRules(projectId));
 
 	function openRuleEditor() {
 		if (editingRuleId !== null) {
@@ -104,13 +52,12 @@
 	}
 
 	function resetEditor() {
-		stackTargetSelected = undefined;
-		draftRuleFilterInitialValues = {};
-		selectedRuleId = undefined;
+		editorInitialStackTarget = undefined;
+		editorInitialFilterValues = {};
 		editingRuleId = null;
 	}
 
-	function cancelRuleEdition() {
+	function closeEditor() {
 		mode = "list";
 		resetEditor();
 	}
@@ -151,75 +98,17 @@
 			return;
 		}
 
-		selectedRuleId = rule.id;
 		editingRuleId = rule.id;
-		stackTargetSelected = rule.action.subject.subject.target;
-		const initialValues: Partial<RuleFilterMap> = {};
+		editorInitialStackTarget = rule.action.subject.subject.target;
 
+		const initialValues: Partial<RuleFilterMap> = {};
 		for (const filter of rule.filters) {
 			updateInitialValues(filter, initialValues);
 		}
-
-		draftRuleFilterInitialValues = initialValues;
+		editorInitialFilterValues = initialValues;
 
 		mode = "edit";
 	}
-
-	async function saveRule() {
-		// Logic to save the rule
-		// This needs to be writtent like this.
-		// If there's no rule filters editor, it means that the match is all files
-		const ruleFilters = ruleFiltersEditor ? ruleFiltersEditor.getRuleFilters() : [];
-
-		if (ruleFilters === undefined) {
-			chipToasts.error("Invalid rule filters");
-			return;
-		}
-
-		if (!stackTargetSelected) {
-			chipToasts.error("Please select a branch to assign the rule");
-			return;
-		}
-
-		if (selectedRuleId) {
-			await update({
-				projectId,
-				request: {
-					id: selectedRuleId,
-					enabled: null,
-					action: {
-						type: "explicit",
-						subject: {
-							type: "assign",
-							subject: { target: stackTargetSelected },
-						},
-					},
-					filters: ruleFilters,
-					trigger: null,
-				},
-			});
-		} else {
-			await create({
-				projectId,
-				request: {
-					action: {
-						type: "explicit",
-						subject: {
-							type: "assign",
-							subject: { target: stackTargetSelected },
-						},
-					},
-					filters: ruleFilters,
-					trigger: "fileSytemChange",
-				},
-			});
-		}
-
-		mode = "list";
-		resetEditor();
-	}
-
-	const rules = $derived(rulesService.workspaceRules(projectId));
 
 	function separateRules(allRules: WorkspaceRule[]) {
 		const regularRules: WorkspaceRule[] = [];
@@ -294,7 +183,7 @@
 			{#if totalRules > 0}
 				{#if mode === "add"}
 					<div in:slide={{ duration: 150 }}>
-						{@render ruleEditor()}
+						<RuleEditor {projectId} onsave={closeEditor} oncancel={closeEditor} />
 					</div>
 				{/if}
 
@@ -303,7 +192,14 @@
 						{#each regularRules.slice().reverse() as rule (rule.id)}
 							{#if editingRuleId === rule.id}
 								<div in:slide={{ duration: 150 }}>
-									{@render ruleEditor()}
+									<RuleEditor
+										{projectId}
+										ruleId={rule.id}
+										initialStackTarget={editorInitialStackTarget}
+										initialFilterValues={editorInitialFilterValues}
+										onsave={closeEditor}
+										oncancel={closeEditor}
+									/>
 								</div>
 							{:else}
 								<Rule {projectId} {rule} editRule={() => editExistingRule(rule)} />
@@ -316,7 +212,14 @@
 							{#each aiRules.slice().reverse() as rule (rule.id)}
 								{#if editingRuleId === rule.id}
 									<div in:slide={{ duration: 150 }}>
-										{@render ruleEditor()}
+										<RuleEditor
+											{projectId}
+											ruleId={rule.id}
+											initialStackTarget={editorInitialStackTarget}
+											initialFilterValues={editorInitialFilterValues}
+											onsave={closeEditor}
+											oncancel={closeEditor}
+										/>
 									</div>
 								{:else}
 									<Rule {projectId} {rule} editRule={() => editExistingRule(rule)} />
@@ -327,7 +230,7 @@
 				</div>
 			{:else if mode === "add"}
 				<div in:slide={{ duration: 170 }}>
-					{@render ruleEditor()}
+					<RuleEditor {projectId} onsave={closeEditor} oncancel={closeEditor} />
 				</div>
 			{:else}
 				<div class="rules-placeholder">
@@ -347,125 +250,6 @@
 			{/if}
 		{/snippet}
 	</ReduxResult>
-{/snippet}
-
-{#snippet ruleEditor()}
-	{@const stackEntries = stackService.stacks(projectId)}
-	<div class="rules-list__editor-content">
-		<div class="rules-list__action">
-			<h3 class="text-13 text-semibold">Assign to branch</h3>
-			<ReduxResult {projectId} result={stackEntries.result}>
-				{#snippet children(stacks)}
-					{@const stackOptions = [
-						...stacks
-							.map((stack) =>
-								stack.id
-									? {
-											label: getStackName(stack),
-											value: encodeStackTarget({ type: "stackId", subject: stack.id }),
-										}
-									: undefined,
-							)
-							.filter(isDefined),
-						{ separator: true } as const,
-						{
-							label: "Leftmost lane",
-							value: encodeStackTarget({ type: "leftmost" }),
-						},
-						{
-							label: "Rightmost lane",
-							value: encodeStackTarget({ type: "rightmost" }),
-						},
-					]}
-
-					<Select
-						value={encodedStackTarget}
-						options={stackOptions}
-						minHeight={200}
-						flex="1"
-						searchable
-						onselect={(selectedStackTarget) => {
-							stackTargetSelected = decodeStackTarget(selectedStackTarget);
-						}}
-						icon={stackTargetSelected
-							? stackTargetSelected.type === "leftmost"
-								? "leftmost-lane"
-								: stackTargetSelected.type === "rightmost"
-									? "rightmost-lane"
-									: "branch"
-							: "branch"}
-					>
-						{#snippet itemSnippet({ item, highlighted })}
-							<SelectItem
-								selected={item.value ? compareStackTarget(item.value, stackTargetSelected) : false}
-								{highlighted}
-							>
-								{item.label || ""}
-								{#snippet iconSnippet()}
-									{#if item.value}
-										{@const target = decodeStackTarget(item.value)}
-										{#if target.type === "leftmost"}
-											<Icon name="leftmost-lane" />
-										{:else if target.type === "rightmost"}
-											<Icon name="rightmost-lane" />
-										{:else}
-											<Icon name="branch" />
-										{/if}
-									{/if}
-								{/snippet}
-							</SelectItem>
-						{/snippet}
-					</Select>
-				{/snippet}
-			</ReduxResult>
-		</div>
-
-		{#if typedKeys(draftRuleFilterInitialValues).length > 0}
-			<div class="rules-list__filters">
-				<RuleFiltersEditor
-					bind:this={ruleFiltersEditor}
-					{projectId}
-					initialFilterValues={draftRuleFilterInitialValues}
-					addFilter={addDraftRuleFilter}
-					deleteFilter={removeDraftRuleFilter}
-				/>
-			</div>
-		{:else}
-			<div class="rules-list__matches-all">
-				<p class="text-13">Matches all changes</p>
-				<div bind:this={addFilterButton} class="rules-list__add-filter-button text-12">
-					<button type="button" onclick={openAddFilterContextMenu}>
-						<span class="clr-text-1 underline-dotted"> Add filter +</span>
-					</button>
-				</div>
-			</div>
-		{/if}
-
-		<Spacer margin={2} dotted />
-
-		<div class="rules-list__editor-buttons">
-			<Button onclick={cancelRuleEdition} kind="outline">Cancel</Button>
-			<Button
-				onclick={saveRule}
-				kind="solid"
-				wide
-				style="gray"
-				disabled={!canSaveRule}
-				loading={stackEntries.result.isLoading ||
-					creatingRule.current.isLoading ||
-					updatingRule.current.isLoading}>Save rule</Button
-			>
-		</div>
-	</div>
-
-	<NewRuleMenu
-		bind:this={newFilterContextMenu}
-		addedFilterTypes={ruleFilterTypes}
-		trigger={addFilterButton}
-		addFromFilter={(type) => {
-			addDraftRuleFilter(type);
-		}}
-	/>
 {/snippet}
 
 <style lang="postcss">
@@ -498,51 +282,6 @@
 	.rules-section {
 		display: flex;
 		flex-direction: column;
-	}
-
-	.rules-list__editor-content {
-		display: flex;
-		flex-direction: column;
-		padding: 12px;
-		padding-bottom: 16px;
-		gap: 16px;
-		border-bottom: 1px solid var(--clr-border-2);
-	}
-
-	.rules-list__filters {
-		display: flex;
-		flex-direction: column;
-	}
-
-	.rules-list__matches-all {
-		display: flex;
-		align-items: center;
-		justify-content: center;
-		padding: 12px;
-		gap: 8px;
-		border-radius: var(--radius-m);
-		background-color: var(--clr-bg-2);
-
-		& p {
-			color: var(--clr-text-2);
-			text-align: center;
-		}
-	}
-
-	.rules-list__action {
-		display: flex;
-		flex-direction: column;
-		gap: 10px;
-	}
-
-	.rules-list__add-filter-button {
-		color: var(--clr-text-2);
-	}
-
-	.rules-list__editor-buttons {
-		display: flex;
-		justify-content: flex-end;
-		gap: 6px;
 	}
 
 	.rule-skeleton {
